@@ -1,11 +1,9 @@
 apply plugin: 'com.android.application'
 apply plugin: 'org.greenrobot.greendao' // apply plugin
 apply plugin: 'com.jakewharton.butterknife'
-apply plugin: 'io.objectbox'
+//apply plugin: 'io.objectbox'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '29.0.2'
 
 //    dexOptions {
 //        javaMaxHeapSize "4g"
@@ -15,21 +13,29 @@ android {
         applicationId "android_serialport_api.xingbang"
         minSdkVersion 19
         targetSdkVersion 30
-        versionCode 22
-        versionName "3.21"
-
-        multiDexEnabled true
-
+        versionCode 25
+        versionName "3.25"
         ndk {
             //设置支持的SO库架构
             abiFilters 'armeabi' //, 'x86', 'armeabi-v7a', 'x86_64', 'arm64-v8a'
         }
+        multiDexEnabled true //防止方法数量超过65536导致错误
     }
-
+    signingConfigs {
+        release {
+            File strfile = new File("D:/Android Work/xingbang_singe/xingbang.jks")
+            storeFile file(strfile)
+            storePassword "123456"
+            keyPassword "123456"
+            keyAlias "xingbang"
+        }
+    }
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled false//是否代码混淆
+
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            signingConfig signingConfigs.release
         }
     }
 
@@ -70,32 +76,39 @@ android {
     viewBinding {
         enabled = true
     }
+    compileSdkVersion 30
+    buildToolsVersion '30.0.3'
+
+
 }
 
 dependencies {
     implementation fileTree(include: '*.jar', dir: 'libs')
     implementation 'androidx.multidex:multidex:2.0.1'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'com.google.code.gson:gson:2.8.6'
     implementation 'com.squareup.okhttp3:okhttp:3.11.0'
     implementation 'com.github.zcweng:switch-button:0.0.3@aar'
-
     implementation 'com.jakewharton:butterknife:10.2.3'
     annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.3'
-
-
     implementation 'com.tencent.bugly:crashreport_upgrade:1.4.1'
     implementation 'org.litepal.guolindev:core:3.2.1'
     implementation 'com.tencent:mmkv-static:1.1.2'
     implementation 'org.greenrobot:greendao:3.3.0'
     implementation 'com.orhanobut:dialogplus:1.11@aar'
-
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-//    debugImplementation 'com.glance.guolindev:glance:1.0.0-alpha01'
+    implementation 'io.github.yuweiguocn:GreenDaoUpgradeHelper:v2.2.1'
+//    debugImplementation 'com.glance.guolindev:glance:1.1.0-alpha01'
+
+    //androidx项目(UI框架)
+//    implementation 'com.github.xuexiangjys:XUI:1.1.8'
+//    implementation 'com.google.android.material:material:1.4.0'
+//    implementation 'com.github.bumptech.glide:glide:4.12.0'
 }
 
 greendao {
-    schemaVersion 21 //数据库版本号
+    schemaVersion 22 //数据库版本号
     targetGenDir 'src/main/java'
     // 设置DaoMaster、DaoSession、Dao 包名
     daoPackage 'android_serialport_api.xingbang.db.greenDao'
