@@ -42,6 +42,7 @@ import com.scandecode.ScanDecode;
 import com.scandecode.inf.ScanInterface;
 
 import org.apache.commons.lang.StringUtils;
+import org.litepal.LitePal;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -49,6 +50,8 @@ import java.util.Date;
 import java.util.List;
 
 import android_serialport_api.xingbang.Application;
+import android_serialport_api.xingbang.db.DetonatorTypeNew;
+import android_serialport_api.xingbang.db.greenDao.DaoMaster;
 import android_serialport_api.xingbang.db.greenDao.DenatorBaseinfoDao;
 import android_serialport_api.xingbang.db.greenDao.DenatorHis_DetailDao;
 import android_serialport_api.xingbang.db.greenDao.MessageBeanDao;
@@ -78,12 +81,10 @@ import butterknife.OnClick;
 import static android_serialport_api.xingbang.Application.getDaoSession;
 
 /**
- * 扫码注册页面
- *
- * @author zenghp
+ * 单发检测
  */
-@TargetApi(Build.VERSION_CODES.HONEYCOMB)
 public class ReisterMainPage_line extends SerialPortActivity implements LoaderCallbacks<Cursor> {
+
     @BindView(R.id.text_start)
     TextView textStart;
     @BindView(R.id.entBF2Bit_st)
@@ -94,10 +95,10 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     EditText entAT1BitSt;
     @BindView(R.id.entboxNoAndSerial_st)
     EditText entboxNoAndSerialSt;
-    @BindView(R.id.btn_ReisterScanStart_st)//扫描起始位
-            Button btnReisterScanStartSt;
-    @BindView(R.id.btn_ReisterScanStart_ed)//扫描按钮终止位
-            Button btnReisterScanStartEd;
+    @BindView(R.id.btn_ReisterScanStart_st)
+    Button btnReisterScanStartSt;//扫描起始位
+    @BindView(R.id.btn_ReisterScanStart_ed)
+    Button btnReisterScanStartEd;//扫描按钮终止位
     @BindView(R.id.entBF2Bit_ed)
     EditText entBF2BitEd;
     @BindView(R.id.entproduceDate_ed)
@@ -157,7 +158,6 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     private String factoryFeature = null;////厂家特征码
     private String deTypeName = null;//雷管类型名称
     private String deTypeSecond = null;//该类型雷管最大延期值
-    private String scanInfo = "";
     //是否单发注册
     private int isSingleReisher = 0;
     //单发注册
@@ -183,15 +183,12 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     private TextView txt_currentVolt;
     private TextView txt_currentIC;
     private SendOpenPower sendOpenThread;
-    private SendPower sendPowerThread;
     private CloseOpenPower closeOpenThread;
-    private ScanBar scanBarThread;
     private ScanInterface scanDecode;
     private volatile int initCloseCmdReFlag = 0;
     private volatile int initOpenCmdReFlag = 0;
     private volatile int revCloseCmdReFlag = 0;
     private volatile int revOpenCmdReFlag = 0;
-    private volatile int tuichu_Flag = 0;//单发检测是否退出注册模式的标识
     private volatile int zhuce_Flag = 0;//单发检测时发送40的标识
     private From12Reister zhuce_form = null;
     private ProgressDialog builder = null;
@@ -209,10 +206,14 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     private SharedPreferencesHelper sharedPreferencesHelper;
     //这是注册了一个观察者模式
     public static final Uri uri = Uri.parse("content://android_serialport_api.xingbang.denatorBaseinfo");
-    private String denator_Type_isSelected = "";//是否设置雷管最大延时
     private String qiaosi_set = "";//是否检测桥丝
+    private String version = "";//版本
     private List<DenatorBaseinfo> list_data = new ArrayList<>();
-    private ArrayList lg2_yanshi = new ArrayList();
+    private final ArrayList lg2_yanshi = new ArrayList();
+    private int send_13 = 0;
+    private int send_10 = 0;
+    private int send_41 = 0;
+    private int send_40 = 0;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -220,8 +221,9 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         setContentView(R.layout.activity_reister_main_page_line);
         ButterKnife.bind(this);
         SoundPlayUtils.init(this);
+
         sharedPreferencesHelper = new SharedPreferencesHelper(this, getApplicationContext().getPackageName());
-        mMyDatabaseHelper = new DatabaseHelper(this, "denatorSys.db", null, 21);
+        mMyDatabaseHelper = new DatabaseHelper(this, "denatorSys.db", null, 22);
         db = mMyDatabaseHelper.getReadableDatabase();
         //扫描参数设置
         init();
@@ -239,23 +241,20 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         reisterListView.setAdapter(adapter);
 
         getLoaderManager().initLoader(0, null, this);
-        reisterListView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-            @Override
-            public boolean onItemLongClick(AdapterView<?> arg0, View view,
-                                           int arg2, long arg3) {
-                LinearLayout myte = (LinearLayout) view;
-                TextView seralNoTxt = (TextView) myte.getChildAt(0);
-                TextView holeTxt = (TextView) myte.getChildAt(1);
-                TextView delayTxt = (TextView) myte.getChildAt(2);
-                TextView denatorTxt = (TextView) myte.getChildAt(3);
-                String serialNo = seralNoTxt.getText().toString().trim();
-                String holeNo = holeTxt.getText().toString().trim();
-                String denatorNo = denatorTxt.getText().toString().trim();
-                String delaytime = delayTxt.getText().toString().trim();
-                selectDenatorId = serialNo;
-                modifyBlastBaseInfo(serialNo, holeNo, delaytime, denatorNo);//序号,孔号,延时,管壳码
-                return false;
-            }
+        reisterListView.setOnItemLongClickListener((arg0, view, arg2, arg3) -> {
+            LinearLayout myte = (LinearLayout) view;
+            TextView seralNoTxt = (TextView) myte.getChildAt(0);
+            TextView holeTxt = (TextView) myte.getChildAt(1);
+            TextView delayTxt = (TextView) myte.getChildAt(2);
+            TextView denatorTxt = (TextView) myte.getChildAt(3);
+
+            String serialNo = seralNoTxt.getText().toString().trim();
+            String holeNo = holeTxt.getText().toString().trim();
+            String denatorNo = denatorTxt.getText().toString().trim();
+            String delaytime = delayTxt.getText().toString().trim();
+            selectDenatorId = serialNo;
+            modifyBlastBaseInfo(serialNo, holeNo, delaytime, denatorNo);//序号,孔号,延时,管壳码
+            return false;
         });
         this.isSingleReisher = 0;
 
@@ -276,14 +275,20 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
             edit_start_entBF2Bit_st.setText(factoryCode);
         }
 
-
         showDenatorSum();//显示雷管总数
+        Utils.writeRecord("---进入单发注册页面---");
+        if (version.equals("01")) {
+            sendCmd(FourStatusCmd.send46("00", "01"));//20(第一代)
+        } else {
+            sendCmd(FourStatusCmd.send46("00", "02"));//20(第二代)
+        }
     }
 
     private void getUserMessage() {
-        List<MessageBean> message = getDaoSession().getMessageBeanDao().queryBuilder().where(MessageBeanDao.Properties.Id.eq((long)1)).list();
-        if(message.size()>0){
+        List<MessageBean> message = getDaoSession().getMessageBeanDao().queryBuilder().where(MessageBeanDao.Properties.Id.eq((long) 1)).list();
+        if (message.size() > 0) {
             qiaosi_set = message.get(0).getQiaosi_set();
+            version = message.get(0).getVersion();
         }
     }
 
@@ -291,138 +296,125 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         scanDecode = new ScanDecode(this);
         scanDecode.initService("true");//初始化扫描服务
 
-        scanDecode.getBarCode(new ScanInterface.OnScanListener() {
-            @Override
-            public void getBarcode(String data) {
-                // scanInfo = data;
-                if (data.length() == 19) {
-                    Log.e("箱号", "getBarcode: " + data);
-                    addXiangHao(data);
-                }
-                if (sanButtonFlag > 0) {
-                    scanDecode.stopScan();
-                    decodeBar(data);
-                } else {
-                    if (continueScanFlag == 1) {
-                        String barCode = getContinueScanBlastNo(data);
-                        if (barCode == null) return;
+        scanDecode.getBarCode(data -> {
+            if (data.length() == 19) {
+                Log.e("箱号", "getBarcode: " + data);
+                addXiangHao(data);
+            }
+            if (sanButtonFlag > 0) {
+                scanDecode.stopScan();
+                decodeBar(data);
+            } else {
+                if (continueScanFlag == 1) {
+                    String barCode = getContinueScanBlastNo(data);
+                    if (barCode == null) return;
 
-                        if (checkRepeatShellNo(barCode) == 1) {
-                            show_Toast(getResources().getString(R.string.text_error_tip9));
-                            SoundPlayUtils.play(3);
-                            return;
-                        } else {
-                            show_Toast(getResources().getString(R.string.text_error_tip10) + barCode);
-                        }
-                        SoundPlayUtils.play(1);//1,2,3,4
-                        insertSingleDenator(barCode);
+                    if (checkRepeatShellNo(barCode)) {
+                        show_Toast(getResources().getString(R.string.text_error_tip9));
+                        SoundPlayUtils.play(3);
+                        return;
+                    } else {
+                        show_Toast(getResources().getString(R.string.text_error_tip10) + barCode);
                     }
+                    SoundPlayUtils.play(1);//1,2,3,4
+                    insertSingleDenator(barCode);
                 }
             }
         });
     }
 
     private void handler() {
-        mHandler_2 = new Handler() {
-            @SuppressLint("HandlerLeak")
-            @Override
-            public void handleMessage(Message msg) {
-                if (pb_show == 1 && tipDlg != null) tipDlg.show();
-                if (pb_show == 0 && tipDlg != null) tipDlg.dismiss();
-                super.handleMessage(msg);
+        mHandler_2 = new Handler(message -> {
+            if (pb_show == 1 && tipDlg != null) tipDlg.show();
+            if (pb_show == 0 && tipDlg != null) tipDlg.dismiss();
+            return false;
+        });
+        mHandler_tip = new Handler(message -> {
+            if (isCorrectReisterFea == 1) {
+                SoundPlayUtils.play(4);
+                show_Toast(getResources().getString(R.string.text_error_tip1));
+                //"雷管信息有误，管厂码不正确，请检查"
+            } else if (isCorrectReisterFea == 2) {
+                SoundPlayUtils.play(4);
+                show_Toast(getResources().getString(R.string.text_error_tip2));
+            } else if (isCorrectReisterFea == 3) {
+                SoundPlayUtils.play(4);
+                show_Toast("已达到最大延时限制" + maxSecond + "ms");
+            } else if (isCorrectReisterFea == 4) {
+                SoundPlayUtils.play(4);
+                show_Toast_long("与第" + lg_No + "发" + singleShellNo + "重复");
+                int total = showDenatorSum();
+                reisterListView.setSelection(total - Integer.parseInt(lg_No));
+            } else if (isCorrectReisterFea == 6) {
+                SoundPlayUtils.play(4);
+                show_Toast("当前管壳码不等于13位,请检查雷管或系统版本是否符合后,再次注册");
+            } else if (isCorrectReisterFea == 99) {
+                adapter.notifyDataSetChanged();
+            } else {
+                SoundPlayUtils.play(4);
+                show_Toast("注册失败");
             }
-        };
-        mHandler_tip = new Handler() {
-            @SuppressLint("HandlerLeak")
-            @Override
-            public void handleMessage(Message msg) {
-                if (isCorrectReisterFea == 1) {
-                    SoundPlayUtils.play(4);
-                    show_Toast(getResources().getString(R.string.text_error_tip1));
-                    //"雷管信息有误，管厂码不正确，请检查"
-                } else if (isCorrectReisterFea == 2) {
-                    SoundPlayUtils.play(4);
-                    show_Toast(getResources().getString(R.string.text_error_tip2));
-                } else if (isCorrectReisterFea == 3) {
-                    SoundPlayUtils.play(4);
-                    show_Toast("已达到最大延时限制" + maxSecond + "ms");
-                } else if (isCorrectReisterFea == 4) {
-                    SoundPlayUtils.play(4);
-                    show_Toast_long("与第" + lg_No + "发" + singleShellNo + "重复");
-                    int total = showDenatorSum();
-                    reisterListView.setSelection(total - Integer.parseInt(lg_No));
-                } else if (isCorrectReisterFea == 6) {
-                    SoundPlayUtils.play(4);
-                    show_Toast("当前管壳码不等于13位,请检查雷管或系统版本是否符合后,再次注册");
-                } else if (isCorrectReisterFea == 99) {
-                    adapter.notifyDataSetChanged();
-                } else {
-                    SoundPlayUtils.play(4);
-                    show_Toast("注册失败");
-                }
-                isCorrectReisterFea = 0;
-                super.handleMessage(msg);
-            }
-        };
-        mHandler_1 = new Handler() {
-            @Override
-            public void handleMessage(Message msg) {
-                if (tipInfoFlag == 1) {
-                    if (busInfo != null) {
-                        txt_currentVolt.setText(getResources().getString(R.string.text_reister_vol) + busInfo.getBusVoltage() + "V");
-                        txt_currentIC.setText(getResources().getString(R.string.text_reister_ele) + Math.round(busInfo.getBusCurrentIa() * 1000) + "μA");
-                        if (Math.round(busInfo.getBusCurrentIa() * 1000) > 80) {//判断当前电流是否偏大
-                            txt_currentIC.setTextColor(Color.RED);
-                        } else {
-                            txt_currentIC.setTextColor(Color.GREEN);
-                        }
+            isCorrectReisterFea = 0;
+
+            return false;
+        });
+        mHandler_1 = new Handler(message -> {
+            if (tipInfoFlag == 1) {
+                if (busInfo != null) {
+                    txt_currentVolt.setText(getResources().getString(R.string.text_reister_vol) + busInfo.getBusVoltage() + "V");
+                    txt_currentIC.setText(getResources().getString(R.string.text_reister_ele) + busInfo.getBusCurrentIa() + "μA");
+                    if (Math.round(busInfo.getBusCurrentIa()) > 51) {//判断当前电流是否偏大
+                        txt_currentIC.setTextColor(Color.RED);
+                    } else {
+                        txt_currentIC.setTextColor(Color.GREEN);
                     }
                 }
-                if (tipInfoFlag == 2) {//提示已注册多少发
-                    if (busInfo != null) {
-                        byte[] reCmd = FourStatusCmd.setToXbCommon_Power_Status24_1("00", "01");//获取电源信息
-                        sendCmd(reCmd);
-                    }
-                    showDenatorSum();//得到数据的总条数
-                }
-                if (tipInfoFlag == 3) {//未收到关闭电源命令
-                    show_Toast("未收到单片机返回命令");
-                }
-                if (tipInfoFlag == 4) {//未收到打开电源命令
-                    show_Toast(getResources().getString(R.string.text_error_tip6));
-                }
-                if (tipInfoFlag == 5) {//桥丝不正常
-                    show_Toast(getResources().getString(R.string.text_error_tip7));
-                    SoundPlayUtils.play(4);
-                }
-                if (tipInfoFlag == 6) {//桥丝不正常
-                    show_Toast("请先设置延时");
-                }
-                if (tipInfoFlag == 7) {//桥丝不正常
-                    show_Toast("当前注册雷管电流过大,请检查雷管");
-                    SoundPlayUtils.play(4);
-                }
-                if (tipInfoFlag == 8) {//桥丝不正常
-                    show_Toast("当前雷管有异常,请检测后重新注册");
-                    SoundPlayUtils.play(4);
-                }
-                if (tipInfoFlag == 9) {//桥丝不正常
-                    show_Toast("当前雷管读码异常,请检查该雷管编码规则");
-                    SoundPlayUtils.play(4);
-                }
-                if (tipInfoFlag == 88) {//刷新界面
-                    showDenatorSum();
-                    edit_start_entboxNoAndSerial_st.getText().clear();
-                    edit_end_entboxNoAndSerial_ed.getText().clear();//.setText("")
-                }
-                if (tipInfoFlag == 89) {//刷新界面
-                    show_Toast("输入的管壳码重复");
-                    showDenatorSum();
-                    SoundPlayUtils.play(4);
-                }
-                super.handleMessage(msg);
             }
-        };
+            if (tipInfoFlag == 2) {//提示已注册多少发
+                if (busInfo != null) {
+                    byte[] reCmd = FourStatusCmd.setToXbCommon_Power_Status24_1("00", "01");//获取电源信息
+                    sendCmd(reCmd);
+                }
+                showDenatorSum();//得到数据的总条数
+            }
+            if (tipInfoFlag == 3) {//未收到关闭电源命令
+                show_Toast("未收到单片机返回命令");
+            }
+            if (tipInfoFlag == 4) {//未收到打开电源命令
+                show_Toast(getResources().getString(R.string.text_error_tip6));
+            }
+            if (tipInfoFlag == 5) {//桥丝不正常
+                show_Toast(getResources().getString(R.string.text_error_tip7));
+                SoundPlayUtils.play(4);
+            }
+            if (tipInfoFlag == 6) {//桥丝不正常
+                show_Toast("请先设置延时");
+            }
+            if (tipInfoFlag == 7) {//桥丝不正常
+                show_Toast("当前注册雷管电流过大,请检查雷管");
+                SoundPlayUtils.play(4);
+            }
+            if (tipInfoFlag == 8) {//桥丝不正常
+                show_Toast("当前雷管有异常,请检测后重新注册");
+                SoundPlayUtils.play(4);
+            }
+            if (tipInfoFlag == 9) {//桥丝不正常
+                show_Toast("当前雷管读码异常,请检查该雷管编码规则");
+                SoundPlayUtils.play(4);
+            }
+            if (tipInfoFlag == 88) {//刷新界面
+                showDenatorSum();
+                edit_start_entboxNoAndSerial_st.getText().clear();
+                edit_end_entboxNoAndSerial_ed.getText().clear();//.setText("")
+            }
+            if (tipInfoFlag == 89) {//刷新界面
+                show_Toast("输入的管壳码重复");
+                showDenatorSum();
+                SoundPlayUtils.play(4);
+            }
+
+            return false;
+        });
     }
 
     private void btn_onClick() {
@@ -583,7 +575,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
 
     private int showDenatorSum() {
         GreenDaoMaster master = new GreenDaoMaster();
-        List<DenatorBaseinfo> list =master.queryDenatorBaseinfoToStatusCode("02");
+        List<DenatorBaseinfo> list = master.queryDenatorBaseinfoToStatusCode("02");
         txtReisteramount.setText("已注册:" + list.size());
         return list.size();
     }
@@ -629,10 +621,10 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
      */
     private void getFactoryCode() {
         GreenDaoMaster master = new GreenDaoMaster();
-        List<Defactory> list=master.queryDefactoryToIsSelected("是");
-        if(list.size()>0){
+        List<Defactory> list = master.queryDefactoryToIsSelected("是");
+        if (list.size() > 0) {
             factoryCode = list.get(0).getDeEntCode();
-            factoryFeature =list.get(0).getDeFeatureCode();
+            factoryFeature = list.get(0).getDeFeatureCode();
         }
         Log.e("厂家管码", "factoryCode: " + factoryCode);
     }
@@ -642,8 +634,8 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
      */
     private void getFactoryType() {
         GreenDaoMaster master = new GreenDaoMaster();
-        List<Denator_type> list=master.queryDefactoryTypeToIsSelected("是");
-        if(list.size()>0){
+        List<Denator_type> list = master.queryDefactoryTypeToIsSelected("是");
+        if (list.size() > 0) {
             deTypeName = list.get(0).getDeTypeName();
             deTypeSecond = list.get(0).getDeTypeSecond();
         }
@@ -658,27 +650,11 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         try {
             mOutputStream.write(reCmd);
         } catch (IOException e) {
-            // TODO Auto-generated catch block
             e.printStackTrace();
         }
         super.sendInterruptCmd();
     }
 
-    //发送命令
-    public synchronized void sendCmd(byte[] mBuffer) {//0627添加synchronized,尝试加锁
-        if (mSerialPort != null && mOutputStream != null) {
-            try {
-                String str = Utils.bytesToHexFun(mBuffer);
-//                Utils.writeLog("Reister sendTo:" + str);
-                Log.e("发送命令", "sendCmd: " + str);
-                mOutputStream.write(mBuffer);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        } else {
-            return;
-        }
-    }
 
     public void displayInputKeyboard(View v, boolean hasFocus) {
         //获取系统 IMM
@@ -711,10 +687,8 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     @Override
     public Loader<Cursor> onCreateLoader(int id, Bundle args) {
         args = new Bundle();
-        // TODO Auto-generated method stub
         args.putString("key", "1");
-        MyLoad myLoad = new MyLoad(ReisterMainPage_line.this, args);
-        return myLoad;
+        return new MyLoad(this, args);
     }
 
     @Override
@@ -740,21 +714,10 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         sharedPreferencesHelper.put("start", setDelayTimeStartDelaytime.getText().toString());
         scanDecode.onDestroy();//回复初始状态
         Utils.saveFile();//把软存中的数据存入磁盘中
-        loadMoreData_all_lg();
+//        loadMoreData_all_lg();//
 
-
-        // TODO Auto-generated method stub
         if (db != null) db.close();
         scanDecode.stopScan();//停止扫描
-        if (scanBarThread != null) {
-            scanBarThread.exit = true;  // 终止线程thread
-            try {
-                scanBarThread.join();
-            } catch (InterruptedException e) {
-                // TODO Auto-generated catch block
-                e.printStackTrace();
-            }
-        }
         super.onDestroy();
         fixInputMethodManagerLeak(this);
     }
@@ -847,8 +810,6 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
 
     @Override
     public boolean onContextItemSelected(MenuItem item) {
-        // TODO Auto-generated method stub
-        ContextMenuInfo menuInfo = (ContextMenuInfo) item.getMenuInfo();
         AdapterView.AdapterContextMenuInfo info = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
         int id = (int) info.id;//这里的info.id对应的就是数据库中_id的值
         String Temp = "";
@@ -861,10 +822,12 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
                 Utils.saveFile();//把软存中的数据存入磁盘中
                 break;
+
             case 2:
                 this.modifyBlastBaseInfo(id);
                 Temp = "修改";
                 break;
+
             default:
                 break;
         }
@@ -873,12 +836,10 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     }
 
     private void modifyBlastBaseInfo(int id) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(ReisterMainPage_line.this);
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
         builder.setIcon(R.drawable.ic_launcher);
         builder.setTitle("请修改雷管信息");
-        //    通过LayoutInflater来加载一个xml的布局文件作为一个View对象
-        View view = LayoutInflater.from(ReisterMainPage_line.this).inflate(R.layout.blastbasedialog, null);
-        //    设置我们自己定义的布局文件作为弹出框的Content
+        View view = LayoutInflater.from(this).inflate(R.layout.blastbasedialog, null);
         builder.setView(view);
         final EditText username = (EditText) view.findViewById(R.id.blast_shellBlastNo_field);
         final EditText password = (EditText) view.findViewById(R.id.blast_delay_field);
@@ -893,20 +854,14 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
             password.setText(age);
             cursor.close();
         }
-        builder.setPositiveButton(getString(R.string.text_alert_sure), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                String a = username.getText().toString().trim();
-                String b = password.getText().toString().trim();
-                //    将输入的用户名和密码打印出来
-                show_Toast("管壳码: " + a + ", 延时: " + b);
-            }
+        builder.setPositiveButton(getString(R.string.text_alert_sure), (dialog, which) -> {
+            String a = username.getText().toString().trim();
+            String b = password.getText().toString().trim();
+            //    将输入的用户名和密码打印出来
+            show_Toast("管壳码: " + a + ", 延时: " + b);
         });
-        builder.setNegativeButton(getString(R.string.text_alert_cancel), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
+        builder.setNegativeButton(getString(R.string.text_alert_cancel), (dialog, which) -> {
 
-            }
         });
         builder.show();
 
@@ -925,53 +880,43 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         serialNoTxt.setText(serialNo);
         denatorNoTxt.setText(denatorNo);
         delaytimeTxt.setText(delaytime);
-        builder.setPositiveButton(getString(R.string.text_alert_sure), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                //String a = username.getText().toString().trim();
-                String b = delaytimeTxt.getText().toString().trim();
-                if (maxSecond != 0 && Integer.parseInt(b) >= maxSecond) {//
-                    isCorrectReisterFea = 3;
-                    mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
-                    dialog.dismiss();
-                } else if (b.trim().length() < 1 || maxSecond > 0 && Integer.parseInt(b) > maxSecond) {
-                    show_Toast(getString(R.string.text_error_tip37));
-                    dialog.dismiss();
-                } else {
-                    modifyDelayTime(selectDenatorId, b);
-                    getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
-                    //    将输入的用户名和密码打印出来
-                    show_Toast(getString(R.string.text_error_tip38));
-                    dialog.dismiss();
-                }
-            }
-        });
-        builder.setNegativeButton(getString(R.string.text_alert_cancel), new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-
-            }
-        });
-        builder.setNeutralButton("删除", new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
+        builder.setPositiveButton(getString(R.string.text_alert_sure), (dialog, which) -> {
+            //String a = username.getText().toString().trim();
+            String b = delaytimeTxt.getText().toString().trim();
+            if (maxSecond != 0 && Integer.parseInt(b) > maxSecond) {//
+                isCorrectReisterFea = 3;
+                mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
                 dialog.dismiss();
-                runPbDialog();
-                new Thread(new Runnable() {
-                    @Override
-                    public void run() {
-                        String whereClause = "shellBlastNo = ?";
-                        String[] whereArgs = {denatorNo};
-                        db.delete(DatabaseHelper.TABLE_NAME_DENATOBASEINFO, whereClause, whereArgs);
-                        Utils.deleteData(ReisterMainPage_line.this);//重新排序雷管
-                        getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
-                        tipDlg.dismiss();
-                        Utils.saveFile();//把软存中的数据存入磁盘中
-                        pb_show = 1;
-                    }
-                }).start();
-
+            } else if (b.trim().length() < 1 || maxSecond > 0 && Integer.parseInt(b) > maxSecond) {
+                show_Toast(getString(R.string.text_error_tip37));
+                dialog.dismiss();
+            } else {
+                Utils.writeRecord("-单发修改延时:" + "-管壳码:" + denatorNo + "-延时:" + b);
+                modifyDelayTime(selectDenatorId, b);
+                getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
+                //    将输入的用户名和密码打印出来
+                show_Toast(getString(R.string.text_error_tip38));
+                dialog.dismiss();
             }
+        });
+        builder.setNegativeButton(getString(R.string.text_alert_cancel), (dialog, which) -> {
+
+        });
+        builder.setNeutralButton("删除", (dialog, which) -> {
+            dialog.dismiss();
+            runPbDialog();
+            Utils.writeRecord("-单发删除:" + "-删除管壳码:" + denatorNo + "-延时" + delaytime);
+            new Thread(() -> {
+                String whereClause = "shellBlastNo = ?";
+                String[] whereArgs = {denatorNo};
+                db.delete(DatabaseHelper.TABLE_NAME_DENATOBASEINFO, whereClause, whereArgs);
+                Utils.deleteData(ReisterMainPage_line.this);//重新排序雷管
+                getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
+                tipDlg.dismiss();
+                Utils.saveFile();//把软存中的数据存入磁盘中
+                pb_show = 1;
+            }).start();
+
         });
         builder.show();
     }
@@ -992,12 +937,11 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
 
         String fromCommad = Utils.bytesToHexFun(cmdBuf);
         if (completeValidCmd(fromCommad) == 0) {
-            fromCommad = this.revCmd;
-            if (this.afterCmd != null && this.afterCmd.length() > 0) this.revCmd = this.afterCmd;
-            else this.revCmd = "";
+//            fromCommad = this.revCmd;//如果revcmd不清空,会导致收到40,但却变成上次的13,先屏蔽
+//            if (this.afterCmd != null && this.afterCmd.length() > 0) this.revCmd = this.afterCmd;
+//            else this.revCmd = "";
             String realyCmd1 = DefCommand.decodeCommand(fromCommad);
             if ("-1".equals(realyCmd1) || "-2".equals(realyCmd1)) {
-                return;
             } else {
                 String cmd = DefCommand.getCmd(fromCommad);
                 if (cmd != null) {
@@ -1011,7 +955,6 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
      * 关闭守护线程
      */
     private void closeThread() {
-        //Thread_stage_1 ttst_1
         if (closeOpenThread != null) {
             closeOpenThread.exit = true;  // 终止线程thread
             try {
@@ -1021,7 +964,6 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 e.printStackTrace();
             }
         }
-        //closeOpenThread =null;
         if (sendOpenThread != null) {
             sendOpenThread.exit = true;  // 终止线程thread
             try {
@@ -1030,123 +972,116 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 e.printStackTrace();
             }
         }
-        if (this.scanBarThread != null) {
-            scanBarThread.exit = true;  // 终止线程thread
-            try {
-                scanBarThread.join();
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-
     }
 
-    /***
+    //发送命令
+    public synchronized void sendCmd(byte[] mBuffer) {//0627添加synchronized,尝试加锁
+        if (mSerialPort != null && mOutputStream != null) {
+            try {
+                String str = Utils.bytesToHexFun(mBuffer);
+//                Utils.writeLog("Reister sendTo:" + str);
+                Log.e("发送命令", str);
+                if (str.contains("C00010")) {
+                    send_10 = 1;
+                } else if (str.contains("C00041")) {
+                    send_41 = 1;
+                } else if (str.contains("C00013")) {
+                    send_13 = 1;
+                } else if (str.contains("C00040")) {
+                    send_40 = 1;
+                }
+                mOutputStream.write(mBuffer);
+                //实验有没有用
+//                mOutputStream.flush();
+//                mSerialPort.tcflush();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        } else {
+            return;
+        }
+    }
+
+    /**
      * 处理芯片返回命令
      */
     private void doWithReceivData(String cmd, byte[] cmdBuf) {
+        String fromCommad = Utils.bytesToHexFun(cmdBuf);
 
         if (DefCommand.CMD_4_XBSTATUS_2.equals(cmd)) {//41开启总线电源指令
-            sendOpenThread.exit = true;
-            Log.e("是否检测桥丝", "qiaosi_set: " + qiaosi_set);
-            if (qiaosi_set.equals("true")) {
-                byte[] reCmd = OneReisterCmd.setToXbCommon_Reister_Init12_2("00", "01");//10 进入自动注册模式(00不检测01检测)桥丝
-                sendCmd(reCmd);
+            send_41 = 0;
+//            sendOpenThread.exit = true;
+//            Log.e("是否检测桥丝", "qiaosi_set: " + qiaosi_set);
+            if (qiaosi_set.equals("true")) {//10 进入自动注册模式(00不检测01检测)桥丝
+                sendCmd(OneReisterCmd.setToXbCommon_Reister_Init12_2("00", "01"));
             } else {
-                byte[] reCmd = OneReisterCmd.setToXbCommon_Reister_Init12_2("00", "00");//10 进入自动注册模式(00不检测01检测)桥丝
-                sendCmd(reCmd);
+                sendCmd(OneReisterCmd.setToXbCommon_Reister_Init12_2("00", "00"));
             }
 
+
         } else if (DefCommand.CMD_1_REISTER_1.equals(cmd)) {//10 进入自动注册模式
+            send_10 = 0;
             //发送获取电源信息
             byte[] reCmd = FourStatusCmd.setToXbCommon_Power_Status24_1("00", "00");//40
             sendCmd(reCmd);
 
         } else if (DefCommand.CMD_1_REISTER_3.equals(cmd)) {//12 有雷管接入
-            try {
-                Thread.sleep(1000);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
+//            try {
+//                Thread.sleep(500);//
+//            } catch (InterruptedException e) {
+//                e.printStackTrace();
+//            }
             //2  连续发三次询问电流指令
             byte[] reCmd = FourStatusCmd.setToXbCommon_Power_Status24_1("00", "00");//40获取电源信息
             sendCmd(reCmd);
-            zhuce_form = OneReisterCmd.decodeFromReceiveAutoDenatorCommand14("00", cmdBuf, qiaosi_set);//桥丝检测
+//            zhuce_form = OneReisterCmd.decodeFromReceiveAutoDenatorCommand14("00", cmdBuf, qiaosi_set);//桥丝检测
+            zhuce_form = OneReisterCmd.decode14_newXinPian("00", cmdBuf, qiaosi_set);//桥丝检测
             if (qiaosi_set.equals("true") && zhuce_form.getWire().equals("无")) {
                 tipInfoFlag = 5;//提示类型桥丝不正常
                 mHandler_1.sendMessage(mHandler_1.obtainMessage());
+                String detonatorId = Utils.GetShellNoById_newXinPian(zhuce_form.getFacCode(), zhuce_form.getFeature(), zhuce_form.getDenaId());
+                Utils.writeRecord("--单发注册--:管壳码:" + serchShellBlastNo("A6214" + zhuce_form.getDenaId()) + " 芯片码:" + detonatorId + "该雷管桥丝异常");
             }
-
-            tuichu_Flag = 1;
             zhuce_Flag = 1;
 
         } else if (DefCommand.CMD_1_REISTER_4.equals(cmd)) {//13 退出自动注册模式
+            send_13 = 0;
+//            if (initCloseCmdReFlag == 1) {//打开电源
+//                revCloseCmdReFlag = 1;
+//                closeOpenThread.exit = true;
+//                sendOpenThread = new SendOpenPower();
+//                sendOpenThread.start();
+//            }
 
-            if (initCloseCmdReFlag == 1 && tuichu_Flag == 0) {//打开电源
-                revCloseCmdReFlag = 1;
-                closeOpenThread.exit = true;
-                sendOpenThread = new SendOpenPower();
-                sendOpenThread.start();
-            }
-
-            if (tuichu_Flag == 1) {
-                byte[] powerCmd = FourStatusCmd.setToXbCommon_OpenPower_42_2("00");//41 开启总线电源指令
-                sendCmd(powerCmd);
-                tuichu_Flag = 0;
-            }
-
-        } else if (DefCommand.CMD_4_XBSTATUS_1.equals(cmd)) {// 40 总线电流电压
-
-            //更新电流信息//可调电压版本,系数为0.011,不可调为0.006
-            //            Log.e("电流", "总线电流: " + fromData.getBusCurrentIa());
-//            Log.e("电流", "zhuce_Flag: " + zhuce_Flag);
-            busInfo = FourStatusCmd.decodeFromReceiveDataPower24_1("00", cmdBuf);
+        } else if (DefCommand.CMD_4_XBSTATUS_1.equals(cmd)) { //40 总线电流电压
+            send_40 = 0;
+            busInfo = FourStatusCmd.decodeFromReceiveDataPower24_1("00", cmdBuf);//解析 40指令
             tipInfoFlag = 1;
             mHandler_1.sendMessage(mHandler_1.obtainMessage());
 
-            if (zhuce_Flag == 1) {
-                Log.e("当前电流", "电流: " + Math.round(busInfo.getBusCurrentIa() * 1000));
-                if (Math.round(busInfo.getBusCurrentIa() * 1000) > 70) {//判断当前电流是否偏大
+            if (zhuce_Flag == 1) {//多次单发注册后闪退,busInfo.getBusCurrentIa()为空
+                String detonatorId = Utils.GetShellNoById_newXinPian(zhuce_form.getFacCode(), zhuce_form.getFeature(), zhuce_form.getDenaId());
+                if (busInfo.getBusCurrentIa() > 70) {//判断当前电流是否偏大
                     tipInfoFlag = 7;
                     mHandler_1.sendMessage(mHandler_1.obtainMessage());
                     SoundPlayUtils.play(4);
                     zhuce_Flag = 0;
+                    Utils.writeRecord("--单发注册--:管壳码:" + serchShellBlastNo("A6214" + zhuce_form.getDenaId()) + "芯片码" + zhuce_form.getDenaId() + "该雷管电流过大");
                 } else {
                     if (zhuce_form != null) {//管厂码,特征码,雷管id
-                        //将管壳码转换为雷管id
-                        String shellNo = Utils.GetShellNoById(zhuce_form.getFacCode(), zhuce_form.getFeature(), zhuce_form.getDenaId());
-//                        String shellNo = Utils.GetShellNoById_new(zhuce_form.getFacCode(), zhuce_form.getFeature(), zhuce_form.getDenaId());
-                        Log.e("自动注册", "管壳码: " + shellNo);
-                        Log.e("自动注册", "shellNo.length(): " + shellNo.length());
-                        if (shellNo.substring(2, 7).equals("00000")) {
-                            tipInfoFlag = 8;
-                            mHandler_1.sendMessage(mHandler_1.obtainMessage());
-                            return;
-                        }
-                        if (shellNo.length()!=13) {
-                            tipInfoFlag = 9;
-                            mHandler_1.sendMessage(mHandler_1.obtainMessage());
-                            return;
-                        }
-                        insertSingleDenator(shellNo, zhuce_form.getWire());//单发注册
-                        SoundPlayUtils.play(1);
+//                        // 获取 管壳码
+//                        String shellNo = new GreenDaoMaster().getShellNo(detonatorId);
+                        insertSingleDenator(detonatorId, zhuce_form);//单发注册
                         zhuce_Flag = 0;
                     }
+
                 }
             }
-            if (tuichu_Flag == 1) {//&&zhuce_Flag==2
-                Log.e("单发注册", "退出注册模式");
-                byte[] powerCmd = OneReisterCmd.setToXbCommon_Reister_Exit12_4("00");//13 退出注册模式
-                sendCmd(powerCmd);
-            }
-
-        } else {
-
         }
 
     }
 
-    /***
+    /**
      * 单发注册
      */
     private int insertSingleDenator(String shellNo) {
@@ -1172,7 +1107,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
             return -1;
         }
         //检查重复数据
-        if (checkRepeatShellNo(shellNo) == 1) {
+        if (checkRepeatdenatorId(shellNo)) {
             singleShellNo = shellNo;
             isCorrectReisterFea = 4;
             mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
@@ -1212,42 +1147,60 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         return 0;
     }
 
-    /***
+    /**
      * 单发注册(存储桥丝状态)
      */
-    private int insertSingleDenator(String shellNo, String wire) {
-        //管厂码
-        String facCode = Utils.getDetonatorShellToFactoryCodeStr(shellNo);
-        //特征码
-        String facFea = Utils.getDetonatorShellToFeatureStr(shellNo);
+    private int insertSingleDenator(String detonatorId, From12Reister zhuce_form) {
+        // 管厂码
+        String facCode = Utils.getDetonatorShellToFactoryCodeStr(detonatorId);
+        // 特征码
+        String facFea = Utils.getDetonatorShellToFeatureStr(detonatorId);
+
+        //352841778FDE5
+        //A62141778FDE5
+        Log.e("注册", "detonatorId: " + detonatorId);
+        String shellBlastNo = serchShellBlastNo("A6214" + zhuce_form.getDenaId());//只能在21年用
+        if (checkRepeatdenatorId(zhuce_form.getDenaId())) {//判断芯片码
+            isCorrectReisterFea = 4;
+            mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
+            return -1;
+        }
+        if (shellBlastNo.length() == 13 && checkRepeatShellNo(shellBlastNo)) {//判断管壳码
+            isCorrectReisterFea = 4;
+            mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
+            return -1;
+        }
+        if (detonatorId.startsWith("00000", 2)) {
+            tipInfoFlag = 8;
+            mHandler_1.sendMessage(mHandler_1.obtainMessage());
+            return -1;
+        }
+        if (detonatorId.length() != 13) {
+            tipInfoFlag = 9;
+            mHandler_1.sendMessage(mHandler_1.obtainMessage());
+            return -1;
+        }
         if (factoryCode != null && factoryCode.trim().length() > 0 && !factoryCode.contains(facCode)) {
             isCorrectReisterFea = 1;
             mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
             return -1;
         }
-        if (shellNo.length() != 13) {//判断管壳码是否是13位
-            isCorrectReisterFea = 6;
-            mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
-            return -1;
-        }
+
         if (factoryFeature != null && factoryFeature.trim().length() > 0 && !factoryFeature.contains(facFea)) {
             isCorrectReisterFea = 2;
             mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
             return -1;
         }
-        if (checkRepeatShellNo(shellNo) == 1) {
-            singleShellNo = "";
-            singleShellNo = shellNo;
-            isCorrectReisterFea = 4;
-            mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
-            return -1;
-        }
+        Log.e("查询生产数据库查管壳码", "detonatorId: " + detonatorId);
+
         if (setDelayTimeStartDelaytime.getText().length() == 0 && reEtF1.getText().length() == 0 && reEtF2.getText().length() == 0) {
             tipInfoFlag = 6;
             mHandler_1.sendMessage(mHandler_1.obtainMessage());
             Log.e("验证是否输入延时", "tipInfoFlag: ");
             return -1;
         }
+
+
         int index = getEmptyDenator(-1);
         int maxNo = getMaxNumberNo();
         int start = Integer.parseInt(String.valueOf(setDelayTimeStartDelaytime.getText()));//开始延时
@@ -1255,25 +1208,25 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         int f2 = Integer.parseInt(String.valueOf(reEtF2.getText()));//f2延时
         int delay = getMaxDelay(maxNo);//获取最大延时
         if (delay_set.equals("f1")) {
-            if (maxSecond != 0 && delay + f1 >= maxSecond) {//
+            if (maxSecond != 0 && delay + f1 > maxSecond) {
                 isCorrectReisterFea = 3;
                 mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
                 return -1;
             }
         } else if (delay_set.equals("f2")) {
-            if (maxSecond != 0 && delay + f2 >= maxSecond) {//
+            if (maxSecond != 0 && delay + f2 > maxSecond) {
                 isCorrectReisterFea = 3;
                 mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
                 return -1;
             }
         }
 
-        if (maxSecond != 0 && f1 >= maxSecond) {//
+        if (maxSecond != 0 && f1 > maxSecond) {//
             isCorrectReisterFea = 3;
             mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
             return -1;
         }
-        if (maxSecond != 0 && f2 >= maxSecond) {//
+        if (maxSecond != 0 && f2 > maxSecond) {//
             isCorrectReisterFea = 3;
             mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
             return -1;
@@ -1293,43 +1246,32 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         }
         db.beginTransaction();
         try {
+            ContentValues values = new ContentValues();
 
-            if (qiaosi_set.equals("true") && wire.equals("无")) {//桥丝检测不合格
-                ContentValues values = new ContentValues();
-                values.put("blastserial", "0");
-                values.put("sithole", "0");
-                values.put("shellBlastNo", shellNo);
-                values.put("delay", 0);
-                values.put("regdate", Utils.getDateFormatLong(new Date()));
-                values.put("statusCode", "01");
-                values.put("statusName", "未注册");
-                values.put("errorCode", "FF");
-                values.put("errorName", "无桥丝");
-                values.put("wire", wire);//桥丝状态
-                //向数据库插入数据
-                db.insert("denatorBaseinfo_all", null, values);
-                return -1;
-            }
-            if (index < 0 && !wire.equals("无")) {//说明没有空余的序号可用
-                ContentValues values = new ContentValues();
+            if (index < 0 && !zhuce_form.getWire().equals("无")) {//说明没有空余的序号可用
                 maxNo++;
                 values.put("blastserial", maxNo);
                 values.put("sithole", maxNo);
-                values.put("shellBlastNo", shellNo);
+                if (shellBlastNo.length() == 13) {
+                    values.put("shellBlastNo", shellBlastNo);
+                } else {
+                    values.put("shellBlastNo", "A6214" + zhuce_form.getDenaId());
+                }
+                Utils.writeRecord("--单发注册--" + "注册雷管码:" + shellBlastNo + "芯片码:" + zhuce_form.getDenaId());
+                values.put("denatorId", zhuce_form.getDenaId());
                 values.put("delay", delay);
                 values.put("regdate", Utils.getDateFormatLong(new Date()));
                 values.put("statusCode", "02");
                 values.put("statusName", "已注册");
                 values.put("errorCode", "FF");
                 values.put("errorName", "正常");
-                values.put("wire", wire);//桥丝状态
+                values.put("wire", zhuce_form.getWire());//桥丝状态
                 //向数据库插入数据
                 db.insert("denatorBaseinfo", null, values);
                 db.insert("denatorBaseinfo_all", null, values);
                 getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
             } else {
-                ContentValues values = new ContentValues();
-                values.put("shellBlastNo", shellNo);//key为字段名，value为值
+                values.put("shellBlastNo", detonatorId);//key为字段名，value为值
                 values.put("statusCode", "02");
                 values.put("statusName", "已注册");
                 values.put("errorCode", "FF");
@@ -1343,25 +1285,26 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         }
         isCorrectReisterFea = 99;
         mHandler_tip.sendMessage(mHandler_tip.obtainMessage());
-//        SoundPlayUtils.play(1);
+        SoundPlayUtils.play(1);
+
 //        Utils.saveFile_max();//把数据库中序号最大的一发保存到列表
         return 0;
     }
 
-    /***
+    /**
      * 生成序列号
      */
     private int insertDenator(String prex, int start, int end) {
         if (end < start) return -1;
         if (start < 0 || end > 99999) return -1;
-        String shellNo = "";
+        String shellNo;
         int maxNo = getMaxNumberNo();
         int flag = 0;
         ContentValues values = new ContentValues();
         int reCount = 0;
         for (int i = start; i <= end; i++) {
             shellNo = prex + String.format("%05d", i);
-            if (checkRepeatShellNo(shellNo) == 1) {
+            if (checkRepeatShellNo(shellNo)) {
                 tipInfoFlag = 89;
                 flag = 1;
                 break;
@@ -1466,24 +1409,48 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
      * @param shellBlastNo
      * @return
      */
-    public int checkRepeatShellNo(String shellBlastNo) {
-        String selection = "shellBlastNo = ?"; // 选择条件，给null查询所有
-        String[] selectionArgs = {shellBlastNo + ""};//选择条件参数,会把选择条件中的？替换成这个数组中的值
-        Cursor cursor = db.query("denatorBaseinfo", null, selection, selectionArgs, null, null, null);
-        if (cursor != null) {  //cursor不位空,可以移动到第一行
-            boolean flag = cursor.moveToFirst();
-            if (flag) {
-                lg_No = cursor.getString(1);
-                cursor.close();
-                return 1;
-            } else {
-                cursor.close();
-                return 0;
-            }
+    public boolean checkRepeatdenatorId(String detonatorId) {
+        Log.e("检查重复的数据", "detonatorId: " + detonatorId);
+        GreenDaoMaster master = new GreenDaoMaster();
+        List<DenatorBaseinfo> list_lg = master.checkRepeatdenatorId(detonatorId);
+        if (list_lg.size() > 0) {
+            Log.e("注册", "list_lg: " + list_lg.toString());
+            lg_No = list_lg.get(0).getBlastserial() + "";
+            singleShellNo = list_lg.get(0).getShellBlastNo();
+            return true;
         } else {
-            return 0;
+            return false;
         }
     }
+
+    /**
+     * 检查重复的数据
+     *
+     * @param shellNo
+     * @return
+     */
+    public boolean checkRepeatShellNo(String shellNo) {
+        Log.e("检查重复的数据", "shellNo: " + shellNo);
+        GreenDaoMaster master = new GreenDaoMaster();
+        List<DenatorBaseinfo> list_lg = master.checkRepeatShellNo(shellNo);
+        if (list_lg.size() > 0) {
+            Log.e("注册", "list_lg: " + list_lg.toString());
+            lg_No = list_lg.get(0).getBlastserial() + "";
+            singleShellNo = list_lg.get(0).getShellBlastNo();
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    /**
+     * 查询生产表中对应的管壳码
+     * */
+    private String serchShellBlastNo(String denatorId) {
+        GreenDaoMaster master = new GreenDaoMaster();
+        return master.queryDetonatorTypeNew(denatorId);
+    }
+
 
     /**
      * 获取第一发雷管
@@ -1511,31 +1478,26 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
 
     private void showAlertDialog() {
         AlertDialog dialog = new AlertDialog.Builder(this)
-                .setTitle("注册提示")//设置对话框的标题//"成功起爆"
+                .setTitle("注册提示")
                 .setMessage("注册列表中的雷管已在起爆历史记录里,是否清空列表")//设置对话框的内容"本次任务成功起爆！"
-                //设置对话框的按钮
-                .setNegativeButton("取消", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        dialog.dismiss();
-                    }
-                })
-                .setPositiveButton("确认清空", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        Application.getDaoSession().getDenatorBaseinfoDao().deleteAll();
-                        getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
-                        dialog.dismiss();
-                    }
+                .setNegativeButton("取消", (dialog1, which) -> dialog1.dismiss())
+                .setPositiveButton("确认清空", (dialog12, which) -> {
+                    Application.getDaoSession().getDenatorBaseinfoDao().deleteAll();
+                    getLoaderManager().restartLoader(1, null, ReisterMainPage_line.this);
+                    dialog12.dismiss();
                 }).create();
         Utils.saveFile();//把软存中的数据存入磁盘中
         dialog.show();
     }
 
+    @SuppressLint("NonConstantResourceId")
     @OnClick({R.id.re_btn_f1, R.id.re_btn_f2, R.id.btn_singleReister, R.id.btn_LookHistory, R.id.btn_setdelay, R.id.btn_ReisterScanStart_st, R.id.btn_ReisterScanStart_ed, R.id.btn_return, R.id.btn_inputOk, R.id.re_et_f1, R.id.re_et_f2, R.id.setDelayTime_startDelaytime})
     public void onViewClicked(View view) {
+
         switch (view.getId()) {
+
             case R.id.re_btn_f1:
+
                 hideInputKeyboard();
                 if (reEtF1.getText().toString().equals("")) {
                     show_Toast("当前设置延时为空,请重新设置");
@@ -1543,11 +1505,11 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 }
                 delay_set = "f1";
                 if (maxSecond != 0) {
-                    if (Integer.parseInt(reEtF1.getText().toString()) >= maxSecond) {
+                    if (Integer.parseInt(reEtF1.getText().toString()) > maxSecond) {
                         show_Toast("当前设置延时已超过最大延时" + maxSecond + "ms,请重新设置");
                         return;
                     }
-                    if (Integer.parseInt(setDelayTimeStartDelaytime.getText().toString()) >= maxSecond) {
+                    if (Integer.parseInt(setDelayTimeStartDelaytime.getText().toString()) > maxSecond) {
                         show_Toast("当前开始延时已超过最大延时" + maxSecond + "ms,请重新设置");
                         return;
                     }
@@ -1559,6 +1521,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 reEtF1.clearFocus();
                 reEtF2.clearFocus();
                 break;
+
             case R.id.re_btn_f2:
                 hideInputKeyboard();
                 if (reEtF2.getText().toString().equals("")) {
@@ -1567,11 +1530,11 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 }
                 delay_set = "f2";
                 if (maxSecond != 0) {
-                    if (Integer.parseInt(reEtF2.getText().toString()) >= maxSecond) {
+                    if (Integer.parseInt(reEtF2.getText().toString()) > maxSecond) {
                         show_Toast("当前设置延时已超过最大延时" + maxSecond + "ms,请重新设置");
                         return;
                     }
-                    if (Integer.parseInt(setDelayTimeStartDelaytime.getText().toString()) >= maxSecond) {
+                    if (Integer.parseInt(setDelayTimeStartDelaytime.getText().toString()) > maxSecond) {
                         show_Toast("当前设置延时已超过最大延时" + maxSecond + "ms,请重新设置");
                         return;
                     }
@@ -1583,7 +1546,9 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 reEtF1.clearFocus();
                 reEtF2.clearFocus();
                 break;
+
             case R.id.btn_singleReister:
+
                 if (delay_set.equals("0")) {
                     show_Toast("请设置延时");
                     break;
@@ -1592,7 +1557,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                     show_Toast("有延时为空,请先设置延时");
                     break;
                 }
-                if (isSingleReisher == 0) {
+                if (isSingleReisher == 0 && send_10 == 0 && send_13 == 0 && send_41 == 0 && send_40 == 0) {
                     String shellBlastNo = serchFristLG();
                     int num = serchFristLGINdenatorHis(shellBlastNo);
 
@@ -1600,14 +1565,16 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                         showAlertDialog();
                     }
                     btnInputOk.setEnabled(false);
-                    btnSingleReister.setText(getResources().getString(R.string.text_singleReister_stop));//"停止注册"
+                    btnSingleReister.setText("停止注册");
                     isSingleReisher = 1;
                     closeThread();
-                    closeOpenThread = new CloseOpenPower();
-                    closeOpenThread.start();
-                } else {
+//                    closeOpenThread = new CloseOpenPower();
+//                    closeOpenThread.start();
+                    sendCmd(FourStatusCmd.setToXbCommon_OpenPower_42_2("00"));//41 开启总线电源指令
+
+                } else if (send_10 == 0 && send_13 == 0 && send_41 == 0 && send_40 == 0) {
                     btnInputOk.setEnabled(true);
-                    btnSingleReister.setText(getResources().getString(R.string.text_singleReister));//"单发注册"
+                    btnSingleReister.setText("单发注册");
                     txt_currentVolt.setText("当前电压:");
                     txt_currentIC.setText("当前电流:");
                     txt_currentIC.setTextColor(Color.BLACK);
@@ -1616,23 +1583,25 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                     initOpenCmdReFlag = 0;
                     revCloseCmdReFlag = 0;
                     revOpenCmdReFlag = 0;
-                    tuichu_Flag = 0;
-                    byte[] powerCmd = OneReisterCmd.setToXbCommon_Reister_Exit12_4("00");//13 退出注册模式
-                    sendCmd(powerCmd);
+                    // 13 退出注册模式
+                    sendCmd(OneReisterCmd.setToXbCommon_Reister_Exit12_4("00"));
+                } else {
+                    show_Toast("正在与单片机通讯,请稍等一下再退出注册模式!");
                 }
                 break;
+
             case R.id.btn_LookHistory:
-                String str2 = getString(R.string.text_query_dqlg);//"当前雷管信息"
                 Intent intent2 = new Intent(this, QueryCurrentDetail_All.class);
-                intent2.putExtra("dataSend", str2);
+                intent2.putExtra("dataSend", "当前雷管信息");
                 startActivityForResult(intent2, 1);
                 break;
+
             case R.id.btn_setdelay:
-                String str3 = "设置延时";//"当前雷管信息"
                 Intent intent3 = new Intent(this, SetDelayTime.class);
-                intent3.putExtra("dataSend", str3);
+                intent3.putExtra("dataSend", "设置延时");
                 startActivityForResult(intent3, 1);
                 break;
+
             case R.id.btn_ReisterScanStart_st:
                 hideInputKeyboard();
                 if (continueScanFlag == 0) {
@@ -1644,6 +1613,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 }
                 sanButtonFlag = 1;
                 break;
+
             case R.id.btn_ReisterScanStart_ed:
                 hideInputKeyboard();
                 if (continueScanFlag == 0) {
@@ -1655,6 +1625,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 }
                 sanButtonFlag = 2;
                 break;
+
             case R.id.btn_return:
                 closeThread();
                 Intent intentTemp = new Intent();
@@ -1662,6 +1633,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                 setResult(1, intentTemp);
                 finish();
                 break;
+
             case R.id.btn_inputOk:
                 hideInputKeyboard();
                 String checstr = checkData();
@@ -1676,31 +1648,24 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                     final int end = Integer.parseInt(edsno);
                     pb_show = 1;
                     runPbDialog();
-                    new Thread(new Runnable() {
-                        @Override
-                        public void run() {
-                            insertDenator(prex, start, end);
-                            Log.e("添加码", "prex: " + prex);
-                            Log.e("添加码", "start: " + start);
-                            Log.e("添加码", "end: " + end);
-                        }
+                    new Thread(() -> {
+                        insertDenator(prex, start, end);
+                        Log.e("添加码", "prex: " + prex);
+                        Log.e("添加码", "start: " + start);
+                        Log.e("添加码", "end: " + end);
                     }).start();
                 } else {
                     show_Toast(checstr);
                 }
                 break;
+
             case R.id.re_et_f1:
-                reEtF1.setBackgroundResource(R.drawable.translucent);
-                reEtF2.setBackgroundResource(R.drawable.translucent);
-                break;
             case R.id.re_et_f2:
-                reEtF1.setBackgroundResource(R.drawable.translucent);
-                reEtF2.setBackgroundResource(R.drawable.translucent);
-                break;
             case R.id.setDelayTime_startDelaytime:
                 reEtF1.setBackgroundResource(R.drawable.translucent);
                 reEtF2.setBackgroundResource(R.drawable.translucent);
                 break;
+
         }
     }
 
@@ -1737,26 +1702,6 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         }
     }
 
-    /**
-     * 获取电源信息
-     */
-    private class SendPower extends Thread {
-        public volatile boolean exit = false;
-
-        public void run() {
-            while (!exit) {
-                try {
-                    //发送获取电源信息
-                    byte[] reCmd = FourStatusCmd.setToXbCommon_Power_Status24_1("00", "00");//40
-                    sendCmd(reCmd);
-                    Thread.sleep(50);
-                } catch (InterruptedException e) {
-                    // TODO Auto-generated catch block
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
 
     private class CloseOpenPower extends Thread {
         public volatile boolean exit = false;
@@ -1775,7 +1720,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                         break;
                     }
                     Thread.sleep(100);
-                    if (zeroCount > 80) {
+                    if (zeroCount > 240) {
                         tipInfoFlag = 3;//未收到关闭电源命令
                         mHandler_1.sendMessage(mHandler_1.obtainMessage());
                         exit = true;
@@ -1783,28 +1728,6 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
                     }
                     zeroCount++;
                 } catch (InterruptedException e) {
-                    // TODO Auto-generated catch block
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-
-    private class ScanBar extends Thread {
-        public volatile boolean exit = false;
-
-        public void run() {
-            int zeroCount = 0;
-
-            while (!exit) {
-                try {
-                    scanDecode.starScan();
-                    Thread.sleep(1250);
-                    //break;
-
-                } catch (InterruptedException e) {
-                    // TODO Auto-generated catch block
                     e.printStackTrace();
                 }
             }
@@ -2115,7 +2038,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
         sanButtonFlag = 0;
     }
 
-    /****
+    /**
      * 校验数据
      */
     private String checkData_delay() {
@@ -2174,9 +2097,7 @@ public class ReisterMainPage_line extends SerialPortActivity implements LoaderCa
     private void loadMoreData_all_lg() {
         lg2_yanshi.clear();
         list_data.clear();
-        StringBuffer sb = new StringBuffer();
-        GreenDaoMaster master = new GreenDaoMaster();
-        list_data = master.queryDenatorBaseinfoToStatusCode("02");
+        list_data = new GreenDaoMaster().queryDenatorBaseinfoToStatusCode("02");
         for (int i = 0; i < list_data.size(); i++) {
             lg2_yanshi.add(list_data.get(i).getDelay() + "");
         }
