@@ -87,7 +87,6 @@ public class OneReisterCmd {
 		if(getFromXbCommon_Reister_AutoReDenator13_3(addr,realyCmd1)==0){
 			if(realyCmd1!=null&&realyCmd1.length()==20){
 				String dataHex =  realyCmd1.substring(6, 20);//取得返回数据
-				Log.e("自动返回雷管id", "返回数据dataHex: "+dataHex );
 				From12Reister vo = new From12Reister();
 				String readStatus = dataHex.substring(0,2);//读取状态
 //				Log.e("自动返回雷管id", "读取状态readStatus: "+readStatus );
@@ -120,7 +119,7 @@ public class OneReisterCmd {
 	 * @return
 	 */
 	public static From12Reister decodeFromReceiveAutoDenatorCommand14(String addr , byte[] cmd,String qiaosi_set){
-
+		//C0001208 FF 00 C5F97817 48 35 6EAAC0  新芯片
 		String fromCommad =  Utils.bytesToHexFun(cmd);
 		String realyCmd1 = DefCommand.decodeCommand(fromCommad);
 
@@ -153,9 +152,9 @@ public class OneReisterCmd {
 				String feature = dataHex.substring(12,14);//特征号
 				char c = (char)Integer.parseInt(feature, 16);
 				vo.setFeature(""+c);
-                Log.e("新编码规则", "特征号 feature: "+feature );
+//                Log.e("新编码规则", "特征号 feature: "+feature );
 //                Log.e("自动返回雷管id", "特征号 feature: "+Integer.parseInt(feature, 16) );
-                Log.e("自动返回雷管id", "特征号 c: "+c );
+//                Log.e("自动返回雷管id", "特征号 c: "+c );
 				String facCode = dataHex.substring(14);//管厂码
 				int a =Integer.parseInt(facCode, 16);
 				if(a<10){
@@ -164,13 +163,58 @@ public class OneReisterCmd {
 				}else {
 					vo.setFacCode(""+a);
 				}
-				Log.e("自动返回雷管id", "10进制管厂码 facCode: "+Integer.parseInt(facCode, 16) );
+//				Log.e("自动返回雷管id", "10进制管厂码 facCode: "+Integer.parseInt(facCode, 16) );
 				return vo;
-			}else{
-
 			}
-		}else{
+		}
+		return null;
+	}
 
+	/***
+	 * 解码检测返回的1.3命令，自动返回雷管ID号(桥丝检测)
+	 * @param addr
+	 * @return
+	 */
+	public static From12Reister decode14_newXinPian(String addr , byte[] cmd,String qiaosi_set){
+		//C0001208 FF 00 C5F97817 48 35 6EAAC0  新芯片
+		String fromCommad =  Utils.bytesToHexFun(cmd);
+		String realyCmd1 = DefCommand.decodeCommand(fromCommad);
+
+		if("-1".equals(realyCmd1)||"-2".equals(realyCmd1)){
+			return null;
+		}
+		if(getFromXbCommon_Reister_AutoReDenator13_3(addr,realyCmd1)==0){
+			if(realyCmd1!=null&&realyCmd1.length()==22){
+				String dataHex =  realyCmd1.substring(6, 22);//取得返回数据
+				From12Reister vo = new From12Reister();
+				String readStatus = dataHex.substring(0,2);//读取状态
+				Log.e("自动返回雷管id", "dataHex: "+dataHex );
+				vo.setReadStatus(readStatus);
+				String wire = dataHex.substring(2,4);//桥丝状态
+//				Log.e("自动返回雷管id", "桥丝状态wire: "+wire );
+				if(qiaosi_set.equals("true")){
+					if(wire.equals("01")){//&&readStatus.equals("FF")
+						vo.setWire("有");
+					}else {
+						vo.setWire("无");
+					}
+				}else {
+					vo.setWire("不检测");
+				}
+				String denaId = dataHex.substring(4,12);//雷管id
+				denaId = Utils.swop4ByteOrder(denaId);//字节码换位
+				vo.setDenaId(denaId);
+//				Log.e("自动返回雷管id", "雷管id denaId: "+denaId );
+				String feature = dataHex.substring(12,14);//特征号
+				vo.setFeature(feature);
+//                Log.e("新编码规则", "特征号 feature: "+feature );
+//                Log.e("自动返回雷管id", "特征号 feature: "+Integer.parseInt(feature, 16) );
+//                Log.e("自动返回雷管id", "特征号 c: "+c );
+				String facCode = dataHex.substring(14);//管厂码
+				vo.setFacCode(facCode);
+//				Log.e("自动返回雷管id", "10进制管厂码 facCode: "+Integer.parseInt(facCode, 16) );
+				return vo;
+			}
 		}
 		return null;
 	}
