@@ -17,7 +17,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             + "statusCode text, "//读取状态//00 雷管正常//01 未注册//02 已注册
             + "statusName text, "//状态名称
             + "errorName text, "//错误名称
-            + "errorCode text, "//错误代码
+            + "errorCode text, "//错误代码//通信状态 ,00-与雷管通信失败，01-延期写入不一致，FF-通信成功，AF-未返回命令
             + "authorization text, "//授权
             + "remark text, " //评论
             + "regdate text, " //注册日期
@@ -28,7 +28,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             + "blastserial integer, "//序号
             + "sithole integer, "//孔号
             + "shellBlastNo text, "//管壳码
-            + "denatorId text, "//
+            + "denatorId text, "//芯片码
             + "delay integer, "//延时
             + "statusCode text, "//读取状态
             + "statusName text, "//状态名称
@@ -171,6 +171,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
             + "server_type2 text, "
             + "pro_dwdm text, "
             +"jiance_time text"
+            +"version text"
             + ")";
 
     //用于错误日志表
