@@ -3,11 +3,14 @@ package android_serialport_api.xingbang.db;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 
+import com.github.yuweiguocn.library.greendao.MigrationHelper;
+
 import org.greenrobot.greendao.database.Database;
 
 import android_serialport_api.xingbang.db.greenDao.DaoMaster;
 import android_serialport_api.xingbang.db.greenDao.DenatorBaseinfoDao;
-import android_serialport_api.xingbang.utils.MigrationHelper;
+import android_serialport_api.xingbang.db.greenDao.MessageBeanDao;
+import android_serialport_api.xingbang.db.greenDao.UserMainDao;
 
 /**
  * Created by xingbang on 2021/4/22.
@@ -28,7 +31,20 @@ public class MyOpenHelper extends DaoMaster.OpenHelper {
     @Override
     public void onUpgrade(Database db, int oldVersion, int newVersion) {
         //操作数据库的更新 有几个表升级都可以传入到下面
-        MigrationHelper.getInstance().migrate(db, DenatorBaseinfoDao.class);
+//        MigrationHelper.getInstance().migrate(db, DenatorBaseinfoDao.class);
+        MigrationHelper.migrate(db, new MigrationHelper.ReCreateAllTableListener() {
+
+            @Override
+            public void onCreateAllTables(Database db, boolean ifNotExists) {
+                DaoMaster.createAllTables(db, ifNotExists);
+            }
+
+            @Override
+            public void onDropAllTables(Database db, boolean ifExists) {
+                DaoMaster.dropAllTables(db, ifExists);
+            }
+        }, MessageBeanDao.class);
     }
+
 
 }
