@@ -38,8 +38,8 @@ public class SetVoltageActivity extends SerialPortActivity {
 
     SharedPreferences.Editor edit;
     private Handler Handler_tip = null;//提示信息
-    private boolean send_low=true;
-    private boolean send_high=true;
+    private boolean send_low = true;
+    private boolean send_high = true;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,40 +54,40 @@ public class SetVoltageActivity extends SerialPortActivity {
 
 
         //根据key找值，找不到返回一个默认值
-        String lowVoltage = sp.getString("lowVoltage", "7");
+        String lowVoltage = sp.getString("lowVoltage", "8.5");
         String highVoltage = sp.getString("highVoltage", "16");
         //赋值给控件
         etSetlowVoltage.setText(lowVoltage);
         etSethighVoltage.setText(highVoltage);
-        Handler_tip = new Handler() {
-            @SuppressLint("HandlerLeak")
-            @Override
-            public void handleMessage(Message msg) {
-                super.handleMessage(msg);
-                Bundle b = msg.getData();
-                String shellStr = b.getString("shellStr");
-                Log.e("命令", "msg.arg1: "+msg.arg1 );
-                if (msg.arg1 == 1) {
+        Handler_tip = new Handler(msg -> {
+            Log.e("命令", "msg.arg1: " + msg.arg1);
+            if (msg.arg1 == 1) {
+                if (msg.obj != null) {
+                    show_Toast("设置低压成功,实际低压输出为" + msg.obj + "V");
+                } else {
                     show_Toast("设置低压成功");
-                    send_low=true;
-                    btnLowVoltage.setEnabled(true);
-                    btnHighVoltage.setEnabled(true);
-                    btnLowVoltage.setBackgroundResource(R.drawable.bt_mainpage_style);
-                }else if(msg.arg1 == 2) {
-                    show_Toast("设置高压成功");
-                    send_high=true;
-                    btnLowVoltage.setEnabled(true);
-                    btnHighVoltage.setEnabled(true);
-                    btnHighVoltage.setBackgroundResource(R.drawable.bt_mainpage_style);
-                }else if(msg.arg1 == 3) {
-                    show_Toast("您输入电压不符合规范,请重新输入");
-                }else {
+                }
+                send_low = true;
+                btnLowVoltage.setEnabled(true);
+                btnHighVoltage.setEnabled(true);
+                btnLowVoltage.setBackgroundResource(R.drawable.bt_mainpage_style);
+            } else if (msg.arg1 == 2) {
+                if (msg.obj != null) {
+                    show_Toast("设置高压成功,实际低压输出为" + msg.obj + "V");
+                } else {
                     show_Toast("设置高压成功");
                 }
-
-
+                send_high = true;
+                btnLowVoltage.setEnabled(true);
+                btnHighVoltage.setEnabled(true);
+                btnHighVoltage.setBackgroundResource(R.drawable.bt_mainpage_style);
+            } else if (msg.arg1 == 3) {
+                show_Toast("您输入电压不符合规范,请重新输入");
             }
-        };
+            return false;
+        });
+        Utils.writeRecord("---进入设置电压页面---");
+        Log.e("设置高压", "send_high: " + send_high);
     }
 
     @Override
@@ -108,30 +108,55 @@ public class SetVoltageActivity extends SerialPortActivity {
                 if (cmd != null) {
                     int localSize = fromCommad.length() / 2;
                     byte[] localBuf = Utils.hexStringToBytes(fromCommad);
-                    Log.e("返回命令","fromCommad : "+ fromCommad +"  cmd : "+cmd);
-                    doWithReceivData(cmd, localBuf,localSize);
+                    doWithReceivData(cmd, localBuf, localSize);
                 }
             }
         }
     }
+
     /***
      * 处理芯片返回命令
      */
     private void doWithReceivData(String cmd, byte[] cmdBuf, int size) {
         byte[] locatBuf = new byte[size];
         System.arraycopy(cmdBuf, 0, locatBuf, 0, size);
+        String fromCommad = Utils.bytesToHexFun(locatBuf);
+        String realyCmd1 = DefCommand.decodeCommand(fromCommad);//005B04BD070000
         if (DefCommand.CMD_5_TEST_8.equals(cmd)) {//5B设置低压
             //普通版本 C0005B00D91EC0
-            //调整版本 C0005B04 41 06 00 FF 867F C0
-            String a =FiveTestingCmd.decodeCmd5B("00", locatBuf);
-
+            //调整版本 C0005B04410600FF867FC0
             Message msg = Handler_tip.obtainMessage();
+            if (size > 7) {
+                //芯片设置值(目前为0)
+                String b = realyCmd1.substring(10, 12);
+                String b1 = realyCmd1.substring(12, 14);
+                String strLow = realyCmd1.substring(6, 8);
+                String strHigh = realyCmd1.substring(8, 10);
+                int volthigh = Integer.parseInt(strHigh, 16) * 256;
+                int voltLowInt = Integer.parseInt(strLow, 16);
+                //可调电压版本,系数为0.011,不可调为0.006
+                double voltTotal = (volthigh + voltLowInt) / 4.095 * 3.0 * 0.006;//普通版本
+                msg.obj = Utils.getFloatToFormat((float) voltTotal, 2, 4);
+            }
             msg.arg1 = 1;
             Handler_tip.sendMessage(msg);
-
-            Log.e("命令", "send_low: "+send_low );
-        }else  if(DefCommand.CMD_5_TEST_9.equals(cmd)){//5C设置高压
+            Log.e("命令", "send_low: " + send_low);
+        } else if (DefCommand.CMD_5_TEST_9.equals(cmd)) {//5C设置高压
             Message msg = Handler_tip.obtainMessage();
+            if (size > 7) {
+                //芯片设置值(目前为0)
+                String b = realyCmd1.substring(10, 12);
+                String b1 = realyCmd1.substring(12, 14);
+
+                String strLow = realyCmd1.substring(6, 8);
+                String strHigh = realyCmd1.substring(8, 10);
+                int volthigh = Integer.parseInt(strHigh, 16) * 256;
+                int voltLowInt = Integer.parseInt(strLow, 16);
+                //可调电压版本,系数为0.011,不可调为0.006
+                double voltTotal = (volthigh + voltLowInt) / 4.095 * 3.0 * 0.011;//普通版本
+                msg.obj = Utils.getFloatToFormat((float) voltTotal, 2, 4);
+            }
+
             msg.arg1 = 2;
             Handler_tip.sendMessage(msg);
         }
@@ -151,11 +176,13 @@ public class SetVoltageActivity extends SerialPortActivity {
             }
         }
     }
-    public static boolean isNumber(String str){
-        String reg ="^[0-9]+(.[0-9]+)?$";
+
+    public static boolean isNumber(String str) {
+        String reg = "^[0-9]+(.[0-9]+)?$";
         return str.matches(reg);
 
     }
+
     public final static boolean isNumericT(String str) {
         Pattern pattern = Pattern.compile("^[-\\+]?[.\\d]*$");
         return pattern.matcher(str).matches();
@@ -166,57 +193,59 @@ public class SetVoltageActivity extends SerialPortActivity {
     public void onViewClicked(View view) {
         switch (view.getId()) {
             case R.id.btn_lowVoltage://设置低压
-                if(send_low){
+                if (send_low) {
                     btnLowVoltage.setEnabled(false);
                     btnHighVoltage.setEnabled(false);
                     btnLowVoltage.setBackgroundResource(R.drawable.bt_mainpage_style_2);
-                    String str_lowVoltage=  etSetlowVoltage.getText().toString();
-                    int low = Double.valueOf(Utils.convertToDouble(str_lowVoltage,7.0)*10).intValue();
-                    Log.e("设置电压", "low: "+low );
-                    if(isNumber(str_lowVoltage)&& low> 60 && low< 120){
+                    String str_lowVoltage = etSetlowVoltage.getText().toString();
+                    int low = Double.valueOf(Utils.convertToDouble(str_lowVoltage, 7.0) * 10).intValue();
+                    Log.e("设置电压", "low: " + low);
+                    Utils.writeRecord("-设置低压:"+low);
+                    if (isNumber(str_lowVoltage) && low > 60 && low < 120) {
                         edit.putString("lowVoltage", str_lowVoltage);
                         edit.commit();//点击提交编辑器
                         byte[] delayBye = Utils.shortToByte((short) low);
                         String delayStr = Utils.bytesToHexFun(delayBye);
-                        String b=delayStr.substring(0,2);
-                        String c=delayStr.substring(2);
-                        byte[] reCmd1 = FourStatusCmd.setToXbCommon_SetLowVoltage("00", c+b);//
+                        String b = delayStr.substring(0, 2);
+                        String c = delayStr.substring(2);
+                        byte[] reCmd1 = FourStatusCmd.setToXbCommon_SetLowVoltage("00", c + b);//
                         sendCmd(reCmd1);
-                        send_low=false;
-                        Log.e("设置低压", " c+b: "+ c+b);
-                    }else {
+                        send_low = false;
+                        Log.e("设置低压", " c+b: " + c + b);
+                    } else {
                         Message msg = Handler_tip.obtainMessage();
                         msg.arg1 = 3;
                         Handler_tip.sendMessage(msg);
                     }
-                }else {
+                } else {
                     show_Toast("请等待单片机返回");
                 }
 
 
                 break;
             case R.id.btn_highVoltage://设置高压
-                if(send_high){
-                    String str_highVoltage=  etSethighVoltage.getText().toString();
-                    int high = Double.valueOf(Utils.convertToDouble(str_highVoltage,16.0)*10).intValue();
-                    Log.e("设置电压", "low: "+high );
-                    if(isNumber(str_highVoltage)&&high>100&&high<200){
+                Log.e("设置高压", "send_high: " + send_high);
+                if (send_high) {
+                    String str_highVoltage = etSethighVoltage.getText().toString();
+                    int high = Double.valueOf(Utils.convertToDouble(str_highVoltage, 16.0) * 10).intValue();
+                    Log.e("设置电压", "high: " + high);
+                    Utils.writeRecord("-设置高压:"+high);
+                    if (isNumber(str_highVoltage) && high > 100 && high < 200) {
                         edit.putString("highVoltage", str_highVoltage);
                         edit.commit();//点击提交编辑器
                         byte[] delayBye = Utils.shortToByte((short) high);
                         String delayStr = Utils.bytesToHexFun(delayBye);
-                        String b=delayStr.substring(0,2);
-                        String c=delayStr.substring(2);
-                        byte[] reCmd1 = FourStatusCmd.setToXbCommon_SetHighVoltage("00",  c+b);
-                        sendCmd(reCmd1);
-                        Log.e("设置高压", " c+b: "+ c+b);
-                        send_high=false;
-                    }else {
+                        String b = delayStr.substring(0, 2);
+                        String c = delayStr.substring(2);
+                        sendCmd(FourStatusCmd.setToXbCommon_SetHighVoltage("00", c + b));//5C
+                        Log.e("设置高压", " c+b: " + c + b);
+                        send_high = false;
+                    } else {
                         Message msg = Handler_tip.obtainMessage();
                         msg.arg1 = 3;
                         Handler_tip.sendMessage(msg);
                     }
-                }else {
+                } else {
                     show_Toast("请等待单片机返回");
                 }
 
