@@ -115,7 +115,7 @@ public class XingBangApproveActivity extends BaseActivity {
         ButterKnife.bind(this);
         sharedPreferencesHelper = new SharedPreferencesHelper(XingBangApproveActivity.this, getApplicationContext().getPackageName());
         initTime();
-        mMyDatabaseHelper = new DatabaseHelper(this, "denatorSys.db", null, 21);
+        mMyDatabaseHelper = new DatabaseHelper(this, "denatorSys.db", null, 22);
         db = mMyDatabaseHelper.getWritableDatabase();
         getUserMessage();
         if (equ_no.equals("")) {
